requirements.txt
absl-py==2.2.2
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudscraper==1.2.71
coinmetrics-api-client==2025.4.24.14
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
datasets==3.5.0
dill==0.3.8
dotenv==0.9.9
emoji==2.14.1
exceptiongroup==1.2.2
feedparser==6.0.11
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
google_search_results==2.4.2
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.3.2
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
namex==0.0.9
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
opt_einsum==3.4.0
optree==0.15.0
orjson==3.10.16
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
polars==1.28.1
praw==7.8.1
prawcore==2.4.0
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.31.0
serpapi==0.1.5
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snscrape @ git+https://github.com/JustAnotherArchivist/snscrape.git@614d4c2029a62d348ca56598f87c425966aaec66
sortedcontainers==2.4.0
soupsieve==2.7
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_keras==0.1
termcolor==3.0.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.15.2
typing_extensions==4.13.2
tzdata==2025.2
update-checker==0.18.0
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
